Very simple flask page with login for admin only
This still uses sqlite3 for the db at the moment.
See access/models.py for details on the db schema.
Users with role set to ROLE_ADMIN will also need to
specify a password. No users are set to ROLE_ADMIN by
default. 2 convenience methods are provided for handling
admin users, make_admin(password) and make_user.

For now, this needs to be done from the python interpreter but
the intention is to allow admins to use these functions from
within the web interface.

To make a user an admin:
$ python
>> from access import db
>> from access.models import User
>> u = User.query.filter_by(email='someone@domain.tld').first()
>> u.make_admin('supersecretpassword')
>> db.session.add(u)
>> db.session.commit()

To remove admin status from a user:
$ python
>> from access import db
>> from access.models import User
>> u = User.query.filter_by(email='someone@domain.tld').first()
>> u.make_user()
>> db.session.add(u)
>> db.session.commit()